refactor(nms): Move utils out of fbc_js_core
