+++ nms/app/util/strings.js
