+++ nms/app/views/features/Features.js
@@ -30,7 +30,7 @@ import TableRow from '@material-ui/core/TableRow';
 import axios from 'axios';
 
 import nullthrows from '../../../shared/util/nullthrows';
-import renderList from '../../../fbc_js_core/util/renderList';
+import renderList from '../../../app/util/renderList';
 import {Route, Routes, useNavigate, useParams} from 'react-router-dom';
 import {makeStyles} from '@material-ui/styles';
 import {useEffect, useState} from 'react';
