+++ nms/app/views/traffic/PolicyTracking.js
@@ -22,7 +22,7 @@ import Text from '../../theme/design-system/Text';
 import TypedSelect from '../../components/TypedSelect';
 
 import {AltFormField} from '../../components/FormField';
-import {base64ToHex, decodeBase64} from '../../../fbc_js_core/util/strings';
+import {base64ToHex, decodeBase64} from '../../util/strings';
 import {makeStyles} from '@material-ui/styles';
 import {policyStyles} from './PolicyStyles';
